+++ src/test/resources/sql-schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS `ims`.`customers` (
 CREATE TABLE IF NOT EXISTS `ims`.`items` (
     `itemid` INT(11) NOT NULL AUTO_INCREMENT,
     `item_name` VARCHAR(40) NULL DEFAULT NULL,
-    `value` DECIMAL(3) NULL DEFAULT NULL,
+    `value` FLOAT NULL DEFAULT NULL,
     PRIMARY KEY (`itemid`)
 );
 CREATE TABLE IF NOT EXISTS `ims`.`orders` (
